requirements.txt
setuptools==66.0.0
torch==2.0.0
torchvision==0.15.1
pandas==1.2.4
matplotlib==3.5.0
ruamel.yaml==0.17.4
moviepy==1.0.3
einops==0.3.0
protobuf==3.20.0
gym==0.23.1
gymnasium-robotics==1.2.1
mujoco==2.3.3
scipy==1.8.0
opencv-python==4.7.0.72
numpy==1.21.0
tensorboard
wandb
termcolor